First step. Working for fixed timesteps.
